refactor(debug): decompose HydrationHud to under 600 lines threshold
Closes #1974
